Sync Tab Panel examples to article splits the tab / tabpanel classes to unique files fixes fontawesome icons
