--- 19-tabpanel/a-tabpanel/src/App.css
@@ -1,83 +1,0 @@
-.tab-panel {
- display: flex;
-}
-.tab-panel.top {
- flex-direction: column;
-}
-.tab-panel.bottom {
- flex-direction: column-reverse;
-}
-.tab-panel.left {
- flex-direction: row;
-}
-.tab-panel.right {
- flex-direction: row-reverse;
-}
-
-.tab-strip {
- background-color: #e8e8e8;
- display: flex;
- justify-content: flex-start;
-}
-.tab-strip.indicator {
- justify-content: center;
-}
-.tab-panel.top .tab-strip,
-.tab-panel.bottom .tab-strip {
- flex-direction: row;
-}
-.tab-panel.left .tab-strip,
-.tab-panel.right .tab-strip {
- flex-direction: column;
-}
-
-.tab {
- text-align: center;
- cursor: pointer;
- border: 2px solid transparent;
-}
-.tab:hover {
- background-color: #d8d8d8;
-}
-.tab-panel.top .tab,
-.tab-panel.bottom .tab {
- padding: 12px 6px;
-}
-.tab-panel.left .tab,
-.tab-panel.right .tab {
- padding: 6px 12px;
-}
-.tab-panel.top .tab.active {
- border-bottom-color: #1e8bfb;
-}
-.tab-panel.left .tab.active {
- border-right-color: #1e8bfb;
-}
-.tab-panel.bottom .tab.active {
- border-top-color: #1e8bfb;
-}
-.tab-panel.right .tab.active {
- border-left-color: #1e8bfb;
-}
-.tab.active:hover {
- background-color: transparent;
-}
-.tab .fa {
- margin-right: 8px;
-}
-
-.card-ct {
- flex: 1;
- display: flex;
-}
-
-.card {
- padding: 12px;
- flex: 1 0 auto;
- display: none;
-}
-
-.card.active {
- display: block !important;
-}
-
