+++ 19-tabpanel/a-tabpanel/src/Tab.css
@@ -1,0 +1,34 @@
+.tab {
+  text-align: center;
+  cursor: pointer;
+  border: 2px solid transparent;
+}
+.tab:hover {
+  background-color: #d8d8d8;
+}
+.tab-panel.top .tab,
+.tab-panel.bottom .tab {
+  padding: 12px 6px;
+}
+.tab-panel.left .tab,
+.tab-panel.right .tab {
+  padding: 6px 12px;
+}
+.tab-panel.top .tab.active {
+  border-bottom-color: #1e8bfb;
+}
+.tab-panel.left .tab.active {
+  border-right-color: #1e8bfb;
+}
+.tab-panel.bottom .tab.active {
+  border-top-color: #1e8bfb;
+}
+.tab-panel.right .tab.active {
+  border-left-color: #1e8bfb;
+}
+.tab.active:hover {
+  background-color: transparent;
+}
+.tab .svg-inline--fa {
+  margin-right: 8px;
+}
